Apply /simplify: use the readers and translators the repo already had
Four review agents over the scene-23 commit. What they found, in order of
what it cost.

The manifest re-parse was shipping a NaN. `compileAnisotropyOptions` built
its composition input with `Number.parseFloat` over the C++ it had just
emitted, so scene 23 -- which computes its intensity -- shipped
`"intensity": null`, the only null number across 129 generated manifests. The
reader for this already existed: `staticNumberValue` reads the pinned AST and
returns undefined rather than NaN. Routing through it, plus a
`staticNumberPair` beside `staticVec3Value` for the array shape the pinned
props use, deletes `StaticEvaluator.numberPair`, `vec2Components`, two
context members and a forwarder. A value the scene computes is now absent
from the manifest, and the composition replays the pin's own default.

The mesh precision fix started one layer too low. `SphereOptions`,
`GroundOptions` and `TorusOptions` carried float scalars, so the new double
chain began from an already-rounded input -- scenes 116 and 162 pass 1.6 and
0.45, which float32 cannot hold. They are doubles from the call site down
now, and both scenes measure every pixel exact.

The generation-decided return was a magic comment string threaded through a
`string[]`, and its docstring claimed a propagation the code did not do: a
nested stop was swallowed. It is a typed `{lines, stopped}` now, so a folded
branch propagates it and a run-time branch refuses by name.

Installing material tracking emitted nothing and recorded nothing, which is
not how this repo drops a pinned behaviour. It reaches `material:tracking`
and writes `material-tracking-observers-dropped` into `fidelity.json` with
the validation that covers it.

Smaller: one pinned-package predicate now answers for dispatch, module
resolution and the capture harness alike, with a test tying the harness regex
to the list; `absentOffsetGuardFields` shares its comparison shape with
`guardedField`, which tightens `guardedField` too; Dawn's fallback face and
its cube flip use the shared helper and swap rows in place instead of
allocating a second buffer; the RGBD decode tabulates its gamma expansion
over the 256 bytes it can be given, bit-identical and 8x faster.

Skipped and filed: lowering the three mesh builders through
`PinnedNumericLowerer`, which already derives precision from the pinned AST
and already serves two other lowerers -- the right depth, and a wave of its
own. The three older extension manifests still re-parse emitted C++; the TODO
now names them and the helper.

All 137 scenes pass all five validation stages; 500 unit tests pass.

+++ native/include/bblite/runtime.hpp
@@ -1643,8 +1643,8 @@ struct Scene {
 };
 
 struct GroundOptions {
-    float width = 1.0f;
-    float height = 1.0f;
+    double width = 1.0;
+    double height = 1.0;
     std::uint32_t subdivisions = 1;
     Vec2 uv_scale{1.0f, 1.0f};
 };
@@ -1660,11 +1660,13 @@ struct PlaneOptions {
     float height = 1.0f;
 };
 
+// The pin halves these as JavaScript numbers before its vertex chain rounds,
+// so they are doubles here for the same reason `CameraRecord`'s scalars are.
 struct SphereOptions {
     std::uint32_t segments = 32;
-    float diameter_x = 1.0f;
-    float diameter_y = 1.0f;
-    float diameter_z = 1.0f;
+    double diameter_x = 1.0;
+    double diameter_y = 1.0;
+    double diameter_z = 1.0;
 };
 
 struct SphereMeshData {
@@ -1677,8 +1679,8 @@ struct SphereMeshData {
 };
 
 struct TorusOptions {
-    float diameter = 1.0f;
-    float thickness = 0.5f;
+    double diameter = 1.0;
+    double thickness = 0.5;
     std::uint32_t tessellation = 16;
 };
